Committed pre-working-loop flat-file templating.
 * The flat file template library, which I will wrap to make a module,
   processes a template and "byte compiles" it into an op-tree.  It has three
   basic operations, and nothing else (currently).
    * String::
      A literal string.
    * Value::
      * A variable (named) value.
    * Enumeration::
      * A loop over an array.

I intend to, from this point where most of the opcodes are chosen, exist and
effectively implemented, I would like to re-write the library because "loop" is
currently a tertiary concept in both the structure and the processing code.  I
find this to be somewhat inappropriate, and would actually like to create
something like a procedure in the opcode and let loop represent an execution of
the procedure.  This will make scope (enumeration depth) a first-class citizen
in the structure and allow for greater flexibility and additional features if I
chose to implement them, such as macros within templates.