A self-shaving autodestructing yak
This was an attempt to factor all of the module parsing and loading
behaviour out into a reusable library which the kernel can use for
module loading, but which can also be used for testing as well as can
be used by the kernel itself.

There's a fair amount of additional change associated with it, and it
should probably be tackled by smaller changes. Placing this off in a
separate branch so it can be teased apart at a later time.