Fixed strange stack-frame error
Now, this is a workaround for what was a bug in gawk. Unfortunately
haven't been able to create a minimum reproducable case for this
issue, but before this change we'd get random errors where in eval.c
we'd be trying to inspect an argument which had no value type.

A little time was spent trying to create a minimum reproducable case
for this issue, more will come (later).